proj2/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
kafka-python==2.0.2
pip-chill==1.0.3
sqlparse==0.5.3
asgiref==3.8.1
confluent-kafka==2.7.0
Django==5.1.4
djangorestframework==3.15.2
pip-chill==1.0.3
six==1.17.0
sqlparse==0.5.3
